Generate Random IDs Using Adjectives and Pokemon Names

Creates unique identifiers by combining random adjectives with Pokemon names in different formats.

Usage

dexid(
  n,
  gen = 1:9,
  case = c("snake", "kebab", "small_camel", "big_camel", "screaming_snake", "parsed",
    "mixed", "lower_upper", "upper_lower", "swap", "all_caps", "lower_camel",
    "upper_camel", "internal_parsing", "none", "flip", "sentence", "random", "title"),
  add_digits = FALSE,
  ...
)

Arguments

n

An integer. Number of IDs to generate.

gen

vector of integers. The generations of Pokemon to use. Defaults to all generations.

case

Character string. The case style to use. Supports all styles implemented in the package snakecase and kebab-case.

add_digits

Logical. If TRUE, adds a random digit to the end of each ID. Defaults to FALSE.

...

other arguments passed to snakecase::to_any_case().

Value

A character vector of length n containing the generated IDs.

Examples

# Generate 5 basic snake_case IDs
dexid(5)
#> [1] "complete_inkay"    "unlined_chewtle"   "advanced_snivy"   
#> [4] "strong_ledyba"     "competent_linoone"

# Generate 5 Title Case IDs
dexid(5, case = "title")
#> [1] "Long Gyarados"           "Mortified Stonjourner"  
#> [3] "Pricey Reshiram"         "Intrepid Drapion"       
#> [5] "Self Assured Wishiwashi"

# Generate 5 IDs with random digits
dexid(5, add_digits = TRUE)
#> [1] "incomparable_quilava_0501" "mundane_golbat_0772"      
#> [3] "quarterly_exeggcute_5606"  "silver_buzzwole_8139"     
#> [5] "unsteady_thievul_0752"    

# Generate 5 IDs restrict to specific generation(s)
dexid(5, gen = 1)
#> [1] "practical_nidoran"   "acrobatic_golduck"   "shady_metapod"      
#> [4] "infinite_clefable"   "scientific_squirtle"